Add files via upload
User is asked if they want a strong password or a regular password. If the user chooses strong, the program will choose 4 random characters out of every symbol, letter (both uppercase and lowercase), and number from 0-9 and places each one in a random index of the password. If the user chooses regular, a random password is generated.